State file fixes
Skip state file writing when there are no source files and the state
file does not already exist.

This fixes an issue where a freezej state file would be written when
slice2freezej was never executed and did not exit. This created a
bogus state file which would fail on the next build.